@@ -1,28 +1,24 @@
 module github.com/AdguardTeam/dnsproxy
 
-go 1.13
+go 1.16
 
 require (
-	github.com/AdguardTeam/golibs v0.4.2
-	github.com/ameshkov/dnscrypt/v2 v2.0.0
+	github.com/AdguardTeam/golibs v0.8.0
+	github.com/ameshkov/dnscrypt/v2 v2.1.3
 	github.com/ameshkov/dnsstamps v1.0.3
-	github.com/beefsack/go-rate v0.0.0-20180408011153-efa7637bb9b6
+	github.com/beefsack/go-rate v0.0.0-20200827232406-6cde80facd47
 	github.com/go-test/deep v1.0.5
-	github.com/golang/protobuf v1.4.3 // indirect
 	github.com/jessevdk/go-flags v1.4.0
-	github.com/joomcode/errorx v1.0.1
+	github.com/joomcode/errorx v1.0.3
 	github.com/kr/text v0.2.0 // indirect
-	github.com/lucas-clemente/quic-go v0.18.1
-	github.com/marten-seemann/qtls-go1-15 v0.1.1 // indirect
-	github.com/miekg/dns v1.1.34
+	github.com/lucas-clemente/quic-go v0.21.1
+	github.com/miekg/dns v1.1.40
 	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
 	github.com/patrickmn/go-cache v2.1.0+incompatible
 	github.com/stretchr/testify v1.6.1
-	golang.org/x/crypto v0.0.0-20201016220609-9e8e0b390897 // indirect
-	golang.org/x/net v0.0.0-20201016165138-7b1cca2348c0
-	golang.org/x/sync v0.0.0-20200317015054-43a5402ce75a // indirect
-	golang.org/x/sys v0.0.0-20201018230417-eeed37f84f13 // indirect
-	google.golang.org/protobuf v1.25.0 // indirect
+	golang.org/x/crypto v0.0.0-20201208171446-5f87f3452ae9 // indirect
+	golang.org/x/net v0.0.0-20210405180319-a5a99cb37ef4
+	golang.org/x/text v0.3.4 // indirect
 	gopkg.in/check.v1 v1.0.0-20200227125254-8fa46927fb4f // indirect
-	gopkg.in/yaml.v3 v3.0.0-20200605160147-a5ece683394c
+	gopkg.in/yaml.v3 v3.0.0-20200615113413-eeeca48fe776
 )
