@@ -1,22 +1,20 @@
 module github.com/lucas-clemente/quic-go
 
-go 1.14
+go 1.15
 
 require (
 	github.com/cheekybits/genny v1.0.0
 	github.com/francoispqt/gojay v1.2.13
-	github.com/golang/groupcache v0.0.0-20191027212112-611e8accdfc9 // indirect
-	github.com/golang/mock v1.4.4
-	github.com/golang/protobuf v1.4.2
-	github.com/marten-seemann/qpack v0.2.0
-	github.com/marten-seemann/qtls v0.10.0
-	github.com/marten-seemann/qtls-go1-15 v0.1.0
+	github.com/golang/mock v1.6.0
+	github.com/marten-seemann/qpack v0.2.1
+	github.com/marten-seemann/qtls-go1-15 v0.1.4
+	github.com/marten-seemann/qtls-go1-16 v0.1.3
+	github.com/marten-seemann/qtls-go1-17 v0.1.0-beta.1.2
 	github.com/onsi/ginkgo v1.14.0
 	github.com/onsi/gomega v1.10.1
-	go.opencensus.io v0.22.2
 	golang.org/x/crypto v0.0.0-20200622213623-75b288015ac9
-	golang.org/x/net v0.0.0-20200707034311-ab3426394381
-	golang.org/x/sync v0.0.0-20190911185100-cd5d95a43a6e
-	google.golang.org/protobuf v1.23.0
+	golang.org/x/net v0.0.0-20210405180319-a5a99cb37ef4
+	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
+	golang.org/x/sys v0.0.0-20210510120138-977fb7262007
 	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
 )
