@@ -14,7 +14,7 @@ import (
 
 // DNSContext represents a DNS request message context
 type DNSContext struct {
-	Proto     string            // "udp", "tcp", "tls", "https", "quic"
+	Proto     Proto
 	Req       *dns.Msg          // DNS request
 	Res       *dns.Msg          // DNS response from an upstream
 	Addr      net.Addr          // client address.
@@ -26,7 +26,8 @@ type DNSContext struct {
 	// If set, Resolve() uses it instead of default servers
 	CustomUpstreamConfig *UpstreamConfig
 
-	// Conn - underlying client connection. Can be null in the case of DOH.
+	// Conn is the underlying client connection.  It is nil if Proto is
+	// ProtoDNSCrypt, ProtoHTTPS, or ProtoQUIC.
 	Conn net.Conn
 
 	// localIP - local IP address (for UDP socket to call udpMakeOOBWithSrc)
@@ -40,19 +41,67 @@ type DNSContext struct {
 	// DNSCryptResponseWriter - necessary to respond to a DNSCrypt query
 	DNSCryptResponseWriter dnscrypt.ResponseWriter
 
-	// QUICStream - QUIC stream from which we got the query (for DOQ only)
+	// QUICStream is the QUIC stream from which we got the query.  For
+	// ProtoQUIC only.
 	QUICStream quic.Stream
+
+	// QUICSession is the QUIC session from which we got the query.  For
+	// ProtoQUIC only.
+	QUICSession quic.Session
+
+	// RequestID is an opaque numerical identifier of this request that is
+	// guaranteed to be unique across requests processed by a single Proxy
+	// instance.
+	RequestID uint64
 
 	ecsReqIP   net.IP // ECS IP used in request
 	ecsReqMask uint8  // ECS mask used in request
+
+	// adBit is the authenticated data flag from the request.
+	adBit bool
+	// hasEDNS0 reflects if the request has EDNS0 RRs.
+	hasEDNS0 bool
+	// doBit is the DNSSEC OK flag from request's EDNS0 RR if presented.
+	doBit bool
+	// udpSize is the UDP buffer size from request's EDNS0 RR if presented,
+	// or default otherwise.
+	udpSize uint16
 }
 
-// scrub - prepares the d.Res to be written (truncates if necessary)
+// calcFlagsAndSize lazily calculates some values required for Resolve method.
+func (ctx *DNSContext) calcFlagsAndSize() {
+	if ctx.udpSize != 0 || ctx.Req == nil {
+		return
+	}
+
+	ctx.adBit = ctx.Req.AuthenticatedData
+	ctx.udpSize = defaultUDPBufSize
+	if o := ctx.Req.IsEdns0(); o != nil {
+		ctx.hasEDNS0 = true
+		ctx.doBit = o.Do()
+		ctx.udpSize = o.UDPSize()
+	}
+}
+
+// scrub prepares the d.Res to be written.  Truncation is applied as well if
+// necessary.
 func (ctx *DNSContext) scrub() {
 	if ctx.Res == nil || ctx.Req == nil {
 		return
 	}
 
-	ctx.Res.Truncate(proxyutil.DNSSize(ctx.Proto, ctx.Req))
-	ctx.Res.Compress = true // some devices require DNS message compression
+	// We should guarantee that all the values we need are calculated.
+	ctx.calcFlagsAndSize()
+
+	// RFC-6891 (https://tools.ietf.org/html/rfc6891) states that response
+	// mustn't contain an EDNS0 RR if the request doesn't include it.
+	//
+	// See https://github.com/AdguardTeam/dnsproxy/issues/132.
+	if ctx.hasEDNS0 && ctx.Res.IsEdns0() == nil {
+		ctx.Res.SetEdns0(ctx.udpSize, ctx.doBit)
+	}
+
+	ctx.Res.Truncate(proxyutil.DNSSize(ctx.Proto == ProtoUDP, ctx.Req))
+	// Some devices require DNS message compression.
+	ctx.Res.Compress = true
 }
