@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"strings"
+	"time"
 
 	"github.com/AdguardTeam/golibs/log"
 	"github.com/joomcode/errorx"
@@ -13,16 +14,21 @@ import (
 
 // NextProtoDQ - During connection establishment, DNS/QUIC support is indicated
 // by selecting the ALPN token "dq" in the crypto handshake.
-// Current draft version: https://tools.ietf.org/html/draft-ietf-dprive-dnsoquic-00
-const NextProtoDQ = "doq-i00"
+// Current draft version: https://datatracker.ietf.org/doc/html/draft-ietf-dprive-dnsoquic-02
+const NextProtoDQ = "doq-i02"
+
+// maxQuicIdleTimeout - maximum QUIC idle timeout.
+// Default value in quic-go is 30, but our internal tests show that
+// a higher value works better for clients written with ngtcp2
+const maxQuicIdleTimeout = 5 * time.Minute
 
 // compatProtoDQ - ALPNs for backwards compatibility
-var compatProtoDQ = []string{"dq", "doq"}
+var compatProtoDQ = []string{"doq-i00", "dq", "doq"}
 
 func (p *Proxy) createQUICListeners() error {
 	for _, a := range p.QUICListenAddr {
 		log.Info("Creating a QUIC listener")
-		quicListen, err := quic.ListenAddr(a.String(), p.TLSConfig, nil)
+		quicListen, err := quic.ListenAddr(a.String(), p.TLSConfig, &quic.Config{MaxIdleTimeout: maxQuicIdleTimeout})
 		if err != nil {
 			return errorx.Decorate(err, "could not start QUIC listener")
 		}
@@ -91,13 +97,8 @@ func (p *Proxy) handleQUICSession(session quic.Session, requestGoroutinesSema se
 // handleQUICStream reads DNS queries from the stream, processes them,
 // and writes back the responses
 func (p *Proxy) handleQUICStream(stream quic.Stream, session quic.Session) {
-	var buf []byte
-	buf = p.bytesPool.Get().([]byte)
-
-	// Linter says that the argument needs to be pointer-like
-	// But it's already pointer-like
-	// nolint
-	defer p.bytesPool.Put(buf)
+	bufPtr := p.bytesPool.Get().(*[]byte)
+	defer p.bytesPool.Put(bufPtr)
 
 	// One query -- one stream
 	// The client MUST send the DNS query over the selected stream, and MUST
@@ -106,6 +107,7 @@ func (p *Proxy) handleQUICStream(stream quic.Stream, session quic.Session) {
 
 	// err is not checked here because STREAM FIN sent by the client is indicated as error here.
 	// instead, we should check the number of bytes received.
+	buf := *bufPtr
 	n, err := stream.Read(buf)
 
 	// The server MUST send the response on the same stream, and MUST indicate through
@@ -125,18 +127,34 @@ func (p *Proxy) handleQUICStream(stream quic.Stream, session quic.Session) {
 		return
 	}
 
-	msg := dns.Msg{}
-	err = msg.Unpack(buf)
+	req := &dns.Msg{}
+	err = req.Unpack(buf)
 	if err != nil {
 		log.Info("failed to unpack a DNS query: %v", err)
 	}
 
-	d := &DNSContext{
-		Proto:      ProtoQUIC,
-		Req:        &msg,
-		Addr:       session.RemoteAddr(),
-		QUICStream: stream,
+	// If any message sent on a DoQ connection contains an edns-tcp-keepalive EDNS(0) Option,
+	// this is a fatal error and the recipient of the defective message MUST forcibly abort
+	// the connection immediately.
+	// https://datatracker.ietf.org/doc/html/draft-ietf-dprive-dnsoquic-02#section-6.6.2
+	if opt := req.IsEdns0(); opt != nil {
+		for _, option := range opt.Option {
+			// Check for EDNS TCP keepalive option
+			if option.Option() == dns.EDNS0TCPKEEPALIVE {
+				log.Debug("client sent EDNS0 TCP keepalive option")
+				errorCode := quic.ApplicationErrorCode(quic.ConnectionRefused)
+
+				// Already closing the connection so we don't care about the error.
+				_ = session.CloseWithError(errorCode, "")
+				return
+			}
+		}
 	}
+
+	d := p.newDNSContext(ProtoQUIC, req)
+	d.Addr = session.RemoteAddr()
+	d.QUICStream = stream
+	d.QUICSession = session
 
 	err = p.handleDNSRequest(d)
 	if err != nil {
@@ -147,6 +165,12 @@ func (p *Proxy) handleQUICStream(stream quic.Stream, session quic.Session) {
 // Writes a response to the QUIC stream
 func (p *Proxy) respondQUIC(d *DNSContext) error {
 	resp := d.Res
+
+	if resp == nil {
+		// If no response has been written, close the QUIC session right away.
+		errorCode := quic.ApplicationErrorCode(quic.InternalError)
+		return d.QUICSession.CloseWithError(errorCode, "")
+	}
 
 	bytes, err := resp.Pack()
 	if err != nil {
@@ -174,8 +198,7 @@ func isQuicConnClosedErr(err error) bool {
 		return true
 	}
 
-	// NO_ERROR: No recent network activity
-	if strings.Contains(str, "NO_ERROR") {
+	if strings.Contains(str, "No recent network activity") {
 		return true
 	}
 
