@@ -3,7 +3,6 @@ package proxy
 import (
 	"fmt"
 	"strings"
-	"time"
 
 	"github.com/AdguardTeam/golibs/log"
 	"github.com/AdguardTeam/golibs/utils"
@@ -25,12 +24,16 @@ type UpstreamConfig struct {
 // So the following config: ["[/host.com/]1.2.3.4", "[/www.host.com/]2.3.4.5", "[/maps.host.com/]#", "3.4.5.6"]
 // will send queries for *.host.com to 1.2.3.4, except for *.www.host.com, which will go to 2.3.4.5 and *.maps.host.com,
 // which will go to default server 3.4.5.6 with all other domains
-func ParseUpstreamsConfig(upstreamConfig, bootstrapDNS []string, timeout time.Duration) (UpstreamConfig, error) {
+func ParseUpstreamsConfig(upstreamConfig []string, options *upstream.Options) (*UpstreamConfig, error) {
+	if options == nil {
+		options = &upstream.Options{}
+	}
+
 	var upstreams []upstream.Upstream
 	domainReservedUpstreams := map[string][]upstream.Upstream{}
 
-	if len(bootstrapDNS) > 0 {
-		log.Debug("Bootstraps: %v", bootstrapDNS)
+	if len(options.Bootstrap) > 0 {
+		log.Debug("Bootstraps: %v", options.Bootstrap)
 	}
 
 	// We use this index to avoid creating duplicates of upstreams
@@ -39,7 +42,7 @@ func ParseUpstreamsConfig(upstreamConfig, bootstrapDNS []string, timeout time.Du
 	for i, l := range upstreamConfig {
 		u, hosts, err := parseUpstreamLine(l)
 		if err != nil {
-			return UpstreamConfig{}, err
+			return &UpstreamConfig{}, err
 		}
 
 		// # excludes more specific domain from reserved upstreams querying
@@ -51,9 +54,16 @@ func ParseUpstreamsConfig(upstreamConfig, bootstrapDNS []string, timeout time.Du
 			dnsUpstream, ok := upstreamsIndex[u]
 			if !ok {
 				// create an upstream
-				dnsUpstream, err = upstream.AddressToUpstream(u, upstream.Options{Bootstrap: bootstrapDNS, Timeout: timeout})
+				dnsUpstream, err = upstream.AddressToUpstream(
+					u,
+					&upstream.Options{
+						Bootstrap:          options.Bootstrap,
+						Timeout:            options.Timeout,
+						InsecureSkipVerify: options.InsecureSkipVerify,
+					})
 				if err != nil {
-					return UpstreamConfig{}, fmt.Errorf("cannot prepare the upstream %s (%s): %s", l, bootstrapDNS, err)
+					err = fmt.Errorf("cannot prepare the upstream %s (%s): %s", l, options.Bootstrap, err)
+					return &UpstreamConfig{}, err
 				}
 
 				// save to the index
@@ -68,14 +78,15 @@ func ParseUpstreamsConfig(upstreamConfig, bootstrapDNS []string, timeout time.Du
 					}
 					domainReservedUpstreams[host] = append(domainReservedUpstreams[host], dnsUpstream)
 				}
-				log.Debug("Upstream %d: %s is reserved for next domains: %s", i, dnsUpstream.Address(), strings.Join(hosts, ", "))
+				log.Debug("Upstream %d: %s is reserved for next domains: %s",
+					i, dnsUpstream.Address(), strings.Join(hosts, ", "))
 			} else {
 				log.Debug("Upstream %d: %s", i, dnsUpstream.Address())
 				upstreams = append(upstreams, dnsUpstream)
 			}
 		}
 	}
-	return UpstreamConfig{
+	return &UpstreamConfig{
 		Upstreams:               upstreams,
 		DomainReservedUpstreams: domainReservedUpstreams,
 	}, nil
@@ -120,26 +131,35 @@ func parseUpstreamLine(l string) (string, []string, error) {
 // If we are looking for domain mail.host.com, this method will return value of host.com key
 // If we are looking for domain www.host.com, this method will return value of www.host.com key
 // If more specific domain value is nil, it means that domain was excluded and should be exchanged with default upstreams
-func (uc *UpstreamConfig) getUpstreamsForDomain(host string) []upstream.Upstream {
+func (uc *UpstreamConfig) getUpstreamsForDomain(host string) (ups []upstream.Upstream) {
 	if len(uc.DomainReservedUpstreams) == 0 {
 		return uc.Upstreams
 	}
 
 	dotsCount := strings.Count(host, ".")
 	if dotsCount < 2 {
-		return uc.DomainReservedUpstreams[UnqualifiedNames]
+		host = UnqualifiedNames
+	} else {
+		host = strings.ToLower(host)
 	}
 
 	for i := 1; i <= dotsCount; i++ {
 		h := strings.SplitAfterN(host, ".", i)
 		name := h[i-1]
-		if u, ok := uc.DomainReservedUpstreams[strings.ToLower(name)]; ok {
-			if u == nil {
-				// domain was excluded from reserved upstreams querying
-				return uc.Upstreams
-			}
-			return u
+
+		var ok bool
+		ups, ok = uc.DomainReservedUpstreams[name]
+		if !ok {
+			continue
 		}
+
+		if len(ups) == 0 {
+			// The domain has been excluded from reserved upstreams
+			// querying.
+			return uc.Upstreams
+		}
+
+		return ups
 	}
 
 	return uc.Upstreams
