@@ -6,12 +6,10 @@ import (
 	"net"
 	"net/url"
 	"strings"
-	"time"
 
 	"github.com/AdguardTeam/dnsproxy/proxyutil"
-	"github.com/miekg/dns"
-
 	"github.com/AdguardTeam/golibs/log"
+	"github.com/miekg/dns"
 )
 
 // Resolver is wrapper for resolver and it's address
@@ -22,9 +20,11 @@ type Resolver struct {
 }
 
 // NewResolver creates an instance of a Resolver structure with defined net.Resolver and it's address
-// resolverAddress is address of net.Resolver
+// resolverAddress -- is address of net.Resolver
 // The host in the address parameter of Dial func will always be a literal IP address (from documentation)
-func NewResolver(resolverAddress string, timeout time.Duration) (*Resolver, error) {
+// options are the upstream customization options, nil means use default
+// options.
+func NewResolver(resolverAddress string, options *Options) (*Resolver, error) {
 	r := &Resolver{}
 
 	// set default net.Resolver as a resolver if resolverAddress is empty
@@ -33,11 +33,16 @@ func NewResolver(resolverAddress string, timeout time.Duration) (*Resolver, erro
 		return r, nil
 	}
 
-	r.resolverAddress = resolverAddress
-	opts := Options{
-		Timeout: timeout,
+	if options == nil {
+		options = &Options{}
 	}
+
+	r.resolverAddress = resolverAddress
 	var err error
+	opts := &Options{
+		Timeout:                 options.Timeout,
+		VerifyServerCertificate: options.VerifyServerCertificate,
+	}
 	r.upstream, err = AddressToUpstream(resolverAddress, opts)
 	if err != nil {
 		log.Error("AddressToUpstream: %s", err)
