@@ -2,6 +2,7 @@
 package upstream
 
 import (
+	"crypto/x509"
 	"fmt"
 	"net"
 	"net/url"
@@ -10,6 +11,7 @@ import (
 	"time"
 
 	"github.com/AdguardTeam/golibs/log"
+	"github.com/ameshkov/dnscrypt/v2"
 	"github.com/ameshkov/dnsstamps"
 	"github.com/joomcode/errorx"
 	"github.com/miekg/dns"
@@ -37,6 +39,13 @@ type Options struct {
 
 	// InsecureSkipVerify - if true, do not verify the server certificate
 	InsecureSkipVerify bool
+
+	// VerifyServerCertificate will be set to crypto/tls Config.VerifyPeerCertificate for DoH, DoQ, DoT
+	VerifyServerCertificate func(rawCerts [][]byte, verifiedChains [][]*x509.Certificate) error
+
+	// VerifyDNSCryptCertificate is callback to which the DNSCrypt server certificate will be passed.
+	// is called in dnsCrypt.exchangeDNSCrypt; if error != nil then Upstream.Exchange() will return it
+	VerifyDNSCryptCertificate func(cert *dnscrypt.Cert) error
 }
 
 // Parse "host:port" string and validate port number
@@ -61,13 +70,18 @@ func parseHostAndPort(addr string) (string, string, error) {
 // * tls://1.1.1.1 -- DNS-over-TLS
 // * https://dns.adguard.com/dns-query -- DNS-over-HTTPS
 // * sdns://... -- DNS stamp (see https://dnscrypt.info/stamps-specifications)
-func AddressToUpstream(address string, opts Options) (Upstream, error) {
+// options -- Upstream customization options, nil means default options.
+func AddressToUpstream(address string, options *Options) (Upstream, error) {
+	if options == nil {
+		options = &Options{}
+	}
+
 	if strings.Contains(address, "://") {
 		upstreamURL, err := url.Parse(address)
 		if err != nil {
 			return nil, errorx.Decorate(err, "failed to parse %s", address)
 		}
-		return urlToUpstream(upstreamURL, opts)
+		return urlToUpstream(upstreamURL, options)
 	}
 
 	// we don't have scheme in the url, so it's just a plain DNS host:port
@@ -79,36 +93,41 @@ func AddressToUpstream(address string, opts Options) (Upstream, error) {
 		port = "53"
 	}
 
-	return &plainDNS{address: net.JoinHostPort(host, port), timeout: opts.Timeout}, nil
+	return &plainDNS{address: net.JoinHostPort(host, port), timeout: options.Timeout}, nil
 }
 
 // urlToBoot creates an instance of the bootstrapper with the specified options
-func urlToBoot(resolverURL string, opts Options) (*bootstrapper, error) {
+// options -- Upstream customization options
+func urlToBoot(resolverURL *url.URL, opts *Options) (*bootstrapper, error) {
 	if len(opts.ServerIPAddrs) == 0 {
-		return newBootstrapper(resolverURL, opts.Bootstrap, opts.Timeout, opts.InsecureSkipVerify)
+		return newBootstrapper(resolverURL, opts)
 	}
 
-	return newBootstrapperResolved(resolverURL, opts.ServerIPAddrs, opts.Timeout, opts.InsecureSkipVerify)
+	return newBootstrapperResolved(resolverURL, opts)
 }
 
 // urlToUpstream converts a URL to an Upstream
-func urlToUpstream(upstreamURL *url.URL, opts Options) (Upstream, error) {
+// options -- Upstream customization options
+func urlToUpstream(upstreamURL *url.URL, opts *Options) (Upstream, error) {
 	switch upstreamURL.Scheme {
 	case "sdns":
-		return stampToUpstream(upstreamURL.String(), opts)
+		return stampToUpstream(upstreamURL, opts)
+
 	case "dns":
 		return &plainDNS{address: getHostWithPort(upstreamURL, "53"), timeout: opts.Timeout}, nil
+
 	case "tcp":
 		return &plainDNS{address: getHostWithPort(upstreamURL, "53"), timeout: opts.Timeout, preferTCP: true}, nil
+
 	case "quic":
 		if upstreamURL.Port() == "" {
-			// https://tools.ietf.org/html/draft-ietf-dprive-dnsoquic-00#section-8.2.1
-			// Early experiments MAY use port 784.  This port is marked in the IANA
-			// registry as unassigned.
-			upstreamURL.Host += ":784"
+			// https://datatracker.ietf.org/doc/html/draft-ietf-dprive-dnsoquic-02#section-10.2.1
+			// Early experiments MAY use port 8853. This port is marked in the IANA registry as unassigned.
+			// (Note that prior to version -02 of this draft, experiments were directed to use port 784.)
+			upstreamURL.Host += ":8853"
 		}
-		resolverURL := upstreamURL.String()
-		b, err := urlToBoot(resolverURL, opts)
+
+		b, err := urlToBoot(upstreamURL, opts)
 		if err != nil {
 			return nil, errorx.Decorate(err, "couldn't create quic bootstrapper")
 		}
@@ -119,8 +138,8 @@ func urlToUpstream(upstreamURL *url.URL, opts Options) (Upstream, error) {
 		if upstreamURL.Port() == "" {
 			upstreamURL.Host += ":853"
 		}
-		resolverURL := upstreamURL.String()
-		b, err := urlToBoot(resolverURL, opts)
+
+		b, err := urlToBoot(upstreamURL, opts)
 		if err != nil {
 			return nil, errorx.Decorate(err, "couldn't create tls bootstrapper")
 		}
@@ -132,8 +151,7 @@ func urlToUpstream(upstreamURL *url.URL, opts Options) (Upstream, error) {
 			upstreamURL.Host += ":443"
 		}
 
-		resolverURL := upstreamURL.String()
-		b, err := urlToBoot(resolverURL, opts)
+		b, err := urlToBoot(upstreamURL, opts)
 		if err != nil {
 			return nil, errorx.Decorate(err, "couldn't create tls bootstrapper")
 		}
@@ -146,10 +164,11 @@ func urlToUpstream(upstreamURL *url.URL, opts Options) (Upstream, error) {
 }
 
 // stampToUpstream converts a DNS stamp to an Upstream
-func stampToUpstream(address string, opts Options) (Upstream, error) {
-	stamp, err := dnsstamps.NewServerStampFromString(address)
+// options -- Upstream customization options
+func stampToUpstream(upsURL *url.URL, opts *Options) (Upstream, error) {
+	stamp, err := dnsstamps.NewServerStampFromString(upsURL.String())
 	if err != nil {
-		return nil, errorx.Decorate(err, "failed to parse %s", address)
+		return nil, errorx.Decorate(err, "failed to parse %s", upsURL)
 	}
 
 	if stamp.ServerAddrStr != "" {
@@ -170,18 +189,20 @@ func stampToUpstream(address string, opts Options) (Upstream, error) {
 	case dnsstamps.StampProtoTypePlain:
 		return &plainDNS{address: stamp.ServerAddrStr, timeout: opts.Timeout}, nil
 	case dnsstamps.StampProtoTypeDNSCrypt:
-		b, err := newBootstrapper(address, opts.Bootstrap, opts.Timeout, opts.InsecureSkipVerify)
+		b, err := newBootstrapper(upsURL, opts)
 		if err != nil {
 			return nil, fmt.Errorf("bootstrap server parse: %s", err)
 		}
 		return &dnsCrypt{boot: b}, nil
 	case dnsstamps.StampProtoTypeDoH:
 		return AddressToUpstream(fmt.Sprintf("https://%s%s", stamp.ProviderName, stamp.Path), opts)
+	case dnsstamps.StampProtoTypeDoQ:
+		return AddressToUpstream(fmt.Sprintf("quic://%s%s", stamp.ProviderName, stamp.Path), opts)
 	case dnsstamps.StampProtoTypeTLS:
 		return AddressToUpstream(fmt.Sprintf("tls://%s", stamp.ProviderName), opts)
 	}
 
-	return nil, fmt.Errorf("unsupported protocol %v in %s", stamp.Proto, address)
+	return nil, fmt.Errorf("unsupported protocol %v in %s", stamp.Proto, upsURL)
 }
 
 // getHostWithPort is a helper function that appends port if needed
