@@ -12,7 +12,7 @@ import (
 func TestUpstreamDNSCrypt(t *testing.T) {
 	// AdGuard DNS (DNSCrypt)
 	address := "sdns://AQIAAAAAAAAAFDE3Ni4xMDMuMTMwLjEzMDo1NDQzINErR_JS3PLCu_iZEIbq95zkSV2LFsigxDIuUso_OQhzIjIuZG5zY3J5cHQuZGVmYXVsdC5uczEuYWRndWFyZC5jb20"
-	u, err := AddressToUpstream(address, Options{Timeout: dialTimeout})
+	u, err := AddressToUpstream(address, &Options{Timeout: dialTimeout})
 	assert.Nil(t, err)
 
 	// Test that it responds properly
@@ -53,7 +53,7 @@ func TestDNSCryptTruncated(t *testing.T) {
 	// Now prepare a client for this test server
 	stamp, err := rc.CreateStamp(udpConn.LocalAddr().String())
 	assert.Nil(t, err)
-	u, err := AddressToUpstream(stamp.String(), Options{Timeout: timeout})
+	u, err := AddressToUpstream(stamp.String(), &Options{Timeout: timeout})
 	assert.Nil(t, err)
 
 	req := new(dns.Msg)
