@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"io/ioutil"
 	"net/http"
+	"os"
 	"sync"
 	"time"
 
@@ -13,25 +14,21 @@ import (
 	"golang.org/x/net/http2"
 )
 
-// DohMaxConnsPerHost controls the maximum number of connections per host
-// nolint
-var DohMaxConnsPerHost = 0
-
-//
-// DNS-over-https
-//
+// dnsOverHTTPS represents DNS-over-HTTPS upstream.
 type dnsOverHTTPS struct {
 	boot *bootstrapper
 
 	// The Client's Transport typically has internal state (cached TCP
 	// connections), so Clients should be reused instead of created as
 	// needed. Clients are safe for concurrent use by multiple goroutines.
-	client *http.Client
-
-	sync.RWMutex // protects transport
+	client      *http.Client
+	clientGuard sync.Mutex
 }
 
-func (p *dnsOverHTTPS) Address() string { return p.boot.address }
+// type check
+var _ Upstream = &dnsOverHTTPS{}
+
+func (p *dnsOverHTTPS) Address() string { return p.boot.URL.String() }
 
 func (p *dnsOverHTTPS) Exchange(m *dns.Msg) (*dns.Msg, error) {
 	client, err := p.getClient()
@@ -42,28 +39,24 @@ func (p *dnsOverHTTPS) Exchange(m *dns.Msg) (*dns.Msg, error) {
 	logBegin(p.Address(), m)
 	r, err := p.exchangeHTTPSClient(m, client)
 	logFinish(p.Address(), err)
-	if err != nil {
-		p.Lock()
-		if client == p.client {
-			p.client = nil // this connection is no longer usable
-		}
-		p.Unlock()
-	}
+
 	return r, err
 }
 
-// exchangeHTTPSClient sends the DNS query to a DOH resolver using the specified http.Client instance
+// exchangeHTTPSClient sends the DNS query to a DOH resolver using the specified
+// http.Client instance.
 func (p *dnsOverHTTPS) exchangeHTTPSClient(m *dns.Msg, client *http.Client) (*dns.Msg, error) {
 	buf, err := m.Pack()
 	if err != nil {
 		return nil, errorx.Decorate(err, "couldn't pack request msg")
 	}
 
-	// It appears, that GET requests are more memory-efficient with Golang implementation of HTTP/2.
-	requestURL := p.boot.address + "?dns=" + base64.RawURLEncoding.EncodeToString(buf)
+	// It appears, that GET requests are more memory-efficient with Golang
+	// implementation of HTTP/2.
+	requestURL := p.Address() + "?dns=" + base64.RawURLEncoding.EncodeToString(buf)
 	req, err := http.NewRequest("GET", requestURL, nil)
 	if err != nil {
-		return nil, errorx.Decorate(err, "couldn't create a HTTP request to %s", p.boot.address)
+		return nil, errorx.Decorate(err, "couldn't create a HTTP request to %s", p.boot.URL)
 	}
 	req.Header.Set("Accept", "application/dns-message")
 
@@ -72,33 +65,43 @@ func (p *dnsOverHTTPS) exchangeHTTPSClient(m *dns.Msg, client *http.Client) (*dn
 		defer resp.Body.Close()
 	}
 	if err != nil {
-		return nil, errorx.Decorate(err, "couldn't do a POST request to '%s'", p.boot.address)
+		// TODO: consider using errors.As
+		if os.IsTimeout(err) {
+			// If this is a timeout error, trying to forcibly re-create the HTTP client instance
+			// See https://github.com/AdguardTeam/AdGuardHome/issues/3217 for more details on this
+			p.clientGuard.Lock()
+			p.client = nil
+			p.clientGuard.Unlock()
+		}
+
+		return nil, errorx.Decorate(err, "couldn't do a GET request to '%s'", p.boot.URL)
 	}
 
 	body, err := ioutil.ReadAll(resp.Body)
 	if err != nil {
-		return nil, errorx.Decorate(err, "couldn't read body contents for '%s'", p.boot.address)
+		return nil, errorx.Decorate(err, "couldn't read body contents for '%s'", p.boot.URL)
 	}
 	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("got an unexpected HTTP status code %d from '%s'", resp.StatusCode, p.boot.address)
+		return nil, fmt.Errorf("got an unexpected HTTP status code %d from '%s'", resp.StatusCode, p.boot.URL)
 	}
 	response := dns.Msg{}
 	err = response.Unpack(body)
 	if err != nil {
-		return nil, errorx.Decorate(err, "couldn't unpack DNS response from '%s': body is %s", p.boot.address, string(body))
+		return nil, errorx.Decorate(err, "couldn't unpack DNS response from '%s': body is %s", p.boot.URL, string(body))
 	}
-	if err == nil && response.Id != m.Id {
+	if response.Id != m.Id {
 		err = dns.ErrId
 	}
 	return &response, err
 }
 
-// getClient gets or lazily initializes an HTTP client (and transport) that will be used for this DOH resolver.
-func (p *dnsOverHTTPS) getClient() (*http.Client, error) {
+// getClient gets or lazily initializes an HTTP client (and transport) that will
+// be used for this DOH resolver.
+func (p *dnsOverHTTPS) getClient() (c *http.Client, err error) {
 	startTime := time.Now()
 
-	p.Lock()
-	defer p.Unlock()
+	p.clientGuard.Lock()
+	defer p.clientGuard.Unlock()
 	if p.client != nil {
 		return p.client, nil
 	}
@@ -106,10 +109,16 @@ func (p *dnsOverHTTPS) getClient() (*http.Client, error) {
 	// Timeout can be exceeded while waiting for the lock
 	// This happens quite often on mobile devices
 	elapsed := time.Since(startTime)
-	if p.boot.timeout > 0 && elapsed > p.boot.timeout {
-		return nil, fmt.Errorf("timeout exceeded: %d ms", int(elapsed/time.Millisecond))
+	if p.boot.options.Timeout > 0 && elapsed > p.boot.options.Timeout {
+		return nil, fmt.Errorf("timeout exceeded: %s", elapsed)
 	}
 
+	p.client, err = p.createClient()
+
+	return p.client, err
+}
+
+func (p *dnsOverHTTPS) createClient() (*http.Client, error) {
 	transport, err := p.createTransport()
 	if err != nil {
 		return nil, errorx.Decorate(err, "couldn't initialize HTTP transport")
@@ -117,45 +126,34 @@ func (p *dnsOverHTTPS) getClient() (*http.Client, error) {
 
 	client := &http.Client{
 		Transport: transport,
-		Timeout:   p.boot.timeout,
+		Timeout:   p.boot.options.Timeout,
 		Jar:       nil,
-	}
-
-	// Warming up the HTTP client.
-	// This is actually important -- if there is no warmup, there's a race condition on the very first DNS query:
-	// http.Client will create numerous connections. During this warmup it'll create a new connection that will be used
-	// for processing further DNS queries.
-	req := dns.Msg{}
-	req.Id = dns.Id()
-	req.RecursionDesired = true
-	req.Question = []dns.Question{{Name: "ipv4only.arpa.", Qtype: dns.TypeA, Qclass: dns.ClassINET}}
-	_, err = p.exchangeHTTPSClient(&req, client)
-	if err != nil {
-		return nil, err
 	}
 
 	p.client = client
 	return p.client, nil
 }
 
-// createTransport initializes an HTTP transport that will be used specifically for this DOH resolver
-// This HTTP transport ensures that the HTTP requests will be sent exactly to the IP address got from the bootstrap resolver
+// createTransport initializes an HTTP transport that will be used specifically
+// for this DOH resolver. This HTTP transport ensures that the HTTP requests
+// will be sent exactly to the IP address got from the bootstrap resolver.
 func (p *dnsOverHTTPS) createTransport() (*http.Transport, error) {
 	tlsConfig, dialContext, err := p.boot.get()
 	if err != nil {
-		return nil, errorx.Decorate(err, "couldn't bootstrap %s", p.boot.address)
+		return nil, errorx.Decorate(err, "couldn't bootstrap %s", p.boot.URL)
 	}
 
 	transport := &http.Transport{
 		TLSClientConfig:    tlsConfig,
 		DisableCompression: true,
 		DialContext:        dialContext,
-		MaxConnsPerHost:    DohMaxConnsPerHost,
-		MaxIdleConns:       1,
 	}
 	// It appears that this is important to explicitly configure transport to use HTTP2
 	// Relevant issue: https://github.com/AdguardTeam/dnsproxy/issues/11
-	http2.ConfigureTransport(transport) // nolint
+	_, err = http2.ConfigureTransports(transport)
+	if err != nil {
+		return nil, err
+	}
 
 	return transport, nil
 }
