@@ -9,8 +9,8 @@ import (
 
 func TestUpstreamDOQ(t *testing.T) {
 	// Create a DNS-over-QUIC upstream
-	address := "quic://dns.adguard.com"
-	u, err := AddressToUpstream(address, Options{InsecureSkipVerify: true})
+	address := "quic://dns.adguard.com:784"
+	u, err := AddressToUpstream(address, &Options{InsecureSkipVerify: true})
 	assert.Nil(t, err)
 
 	uq := u.(*dnsOverQUIC)
